PlmIsing.isingplm — Method

isingplm(spin::Matrix,W::Vector,kwds...)

Pseudo likelihood maximization of an ising pair spin interaction system, from a N × M spin::Matrix where N is number of Ising spins (±1) and M is the number of configurations. W is a vector of M non negative normalized (sum(W) ≈ 1) weights. Keyword args:

* lamdaJ::Real=0.01; lagrange multiplier for the Js

* lambdaH::Real=0.01; lagrange multiplier for the hs

* epsconv::Real::1e-5; convergence paramter

* maxeval::Int=1000; maximal number of iterations

* verbose::Bool=true; verbosity of the output

* method::Symbol=:LD_LBFGS; NLopt maximization strategy

Output: a PlmOut struct containing the Js (as a N × N matrix ), the h (as a length N vector), and the site pseudo-likelihood.

isingplm(filename::String;kwds...)

Read the configurations from a filename. The file is a N × M matrix. In this case the weights are assumed to be all 1.0/M.

PlmIsing Documentation

PlmIsing.isingplm — Function

isingplm(spin::Matrix,W::Vector,kwds...)

Pseudo likelihood maximization of an ising pair spin interaction system, from a N × M spin::Matrix where N is number of Ising spins (±1) and M is the number of configurations. W is a vector of M non negative normalized (sum(W) ≈ 1) weights. Keyword args:

* lamdaJ::Real=0.01; lagrange multiplier for the Js

* lambdaH::Real=0.01; lagrange multiplier for the hs

* epsconv::Real::1e-5; convergence paramter

* maxeval::Int=1000; maximal number of iterations

* verbose::Bool=true; verbosity of the output

* method::Symbol=:LD_LBFGS; NLopt maximization strategy

Output: a PlmOut struct containing the Js (as a N × N matrix ), the h (as a length N vector), and the site pseudo-likelihood.

isingplm(filename::String;kwds...)

Read the configurations from a filename. The file is a N × M matrix. In this case the weights are assumed to be all 1.0/M.